+++ src/main/java/com/powers/PowersBootstrap.java
@@ -10,6 +10,7 @@ import com.powers.item.ArtifactWeaponManager;
 import com.powers.knowledge.KnowledgeEntryReloadListener;
 import com.powers.loot.PowersLoot;
 import com.powers.network.PowersPackets;
+import com.powers.player.PlayerPowers;
 import com.powers.player.SkillSystem;
 import com.powers.power.PowerRegistry;
 import com.powers.power.crystals.CrystalPowerRegistry;
@@ -22,6 +23,7 @@ final class PowersBootstrap {
 	}
 
 	static void initialize() {
+		PlayerPowers.initialize();
 		PowersConfigLoader.initialize();
 		com.powers.command.PermissionNodes.installFabricAdapterIfPresent();
 		KnowledgeEntryReloadListener.initialize();